requirements.txt
streamlit==1.42.0
sqlalchemy==2.0.45
psycopg2-binary==2.9.11
pandas==2.2.2
numpy==2.0.2
python-dotenv==1.0.1